rename output module
