+++ NOTES.md
@@ -13,6 +13,7 @@ TODOs:
   * validate `min/max` etc values cannot be negative, use `u64` instead of `i64`
   * fix debug formatting of serde values is errors - very unreadable right now
   * have a type for `non-empty vec` and use it pretty much everywhere where `Option<Vec>` is used now
+  * naming - is is called `ui object` or `ui schema` ?
    
 * docs
 * after adding support for basic features:do a sweep through the spec and add tests for edge cases etc
