+++ src/lib.rs
@@ -1,4 +1,4 @@
 use serde_derive::Deserialize;
 
 pub mod dsl;
-pub mod ui_configuration;
+pub mod output;
